requirements.txt
amqp==5.3.1
arabic-reshaper==3.0.0
asgiref==3.11.1
asn1crypto==1.5.1
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
billiard==4.2.4
cbor2==5.8.0
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
cryptography==46.0.5
cssselect2==0.9.0
daphne==4.2.1
Django==4.2.28
django-allauth==65.14.3
django-cors-headers==4.9.0
django-extensions==4.1
django-filter==25.1
django-jazzmin==3.0.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
freetype-py==2.5.1
gunicorn==25.1.0
html5lib==1.1
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
kombu==5.6.2
lxml==6.0.2
msgpack==1.1.2
mysqlclient==2.2.8
openpyxl==3.1.5
oscrypto==1.3.0
packaging==26.0
pillow==12.1.1
prompt_toolkit==3.0.52
py-ubjson==0.16.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycairo==1.29.0
pycparser==3.0
pyHanko==0.33.0
pyhanko-certvalidator==0.29.1
PyJWT==2.11.0
pyOpenSSL==25.3.0
pypdf==6.7.2
python-bidi==0.6.7
python-dateutil==2.9.0.post0
PyYAML==6.0.3
redis==7.2.0
reportlab==4.4.10
requests==2.32.5
rlPyCairo==0.4.0
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.5
svglib==1.6.0
tinycss2==1.5.1
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ujson==5.11.0
uritools==6.0.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
webencodings==0.5.1
xhtml2pdf==0.2.17
zope.interface==8.2
